hard code the OGR.OFTInteger(0) to np.int32, OGR.OFTInteger64 (12) to np.int64, and OGR.Real (2) to np.float64

+++ tests/feature/test_feature.py
@@ -73,7 +73,7 @@ class TestAttributes:
         dtypes = feature.dtypes
         assert isinstance(dtypes, dict)
         assert dtypes == {
-            "id": "int64",
+            "id": "int32",
             "x": "float64",
             "y": "float64",
         }
